% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrate.R
\name{data_integrate}
\alias{data_integrate}
\title{Integrate fish data}
\usage{
data_integrate(
  data_path,
  sources = c("Baystudy", "Suisun", "FMWT", "SKT", "DJFMP", "EDSM", "TMM", "SLS", "STN",
    "Salvage"),
  format = "rda",
  write = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data_path}{Path to the folder where you wish the csv/rda files to be saved.}

\item{sources}{Character vector of data sources to include}

\item{format}{Should the fish and survey tables be saved as a csv alone, a combined .rda file, or both? Defaults to 'rda' and other options are 'csv' or 'both'.}

\item{write}{Logical. Should the files be written to disk, or would you just like them returned as an R object? Defaults to TRUE.}

\item{quiet}{Logical. Set to TRUE if you wish to hide all status messages.}
}
\value{
Invisibly returns a list with the survey and fish tables.
}
\description{
Integrate datasets, add zeroes, divide into fish and survey tables, and export to csv and/or rda.
}
\examples{
\dontrun{
data_integrate(data_path=file.path("data-raw", "EDI", "data_objects"))
}
}
