% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{DJFMP}
\alias{DJFMP}
\title{DJFMP dataset}
\format{
a tibble with 2,336,862 rows and 23 variables
\describe{
\item{Source}{Name of source dataset.}
\item{Station}{Station where sample was collected.}
\item{Latitude}{Latitude (decimal degrees).}
\item{Longitude}{Longitude (decimal degrees).}
\item{Date}{Date sample was collected.}
\item{Datetime}{Date and time sample was collected.}
\item{Depth}{Bottom depth (m).}
\item{SampleID}{Unique sample identifier.}
\item{Method}{Sampling method (Otter Trawl or Midwater Trawl).}
\item{Sal_surf}{Surface salinity.}
\item{Temp_surf}{Surface temperature in °C.}
\item{TurbidityNTU}{Turbidity (NTU).}
\item{TurbidityFNU}{Turbidity (FNU).}
\item{Secchi}{Secchi depth (cm).}
\item{Tow_volume}{Volume towed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}}).}
\item{Tow_direction}{Tow direction relative to current.}
\item{Taxa}{Scientific name.}
\item{Length}{Fork length from tip of the snout to a point at the fork of the caudal fin (mm).}
\item{Count}{Estimated count for each sample, taxa, and length.}
\item{Length_NA_flag}{Why is the length NA?}
}
}
\usage{
DJFMP
}
\description{
US Fish and Wildlife Service Delta Juvenile Fish Monitoring Program data.
}
\details{
Some station locations have moved over time due to shoals filling in, trees, etc. More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.244.7}{here}.
}
\keyword{datasets}
