% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{Integrate fish datasets}
\usage{
fish(
  sources,
  species = NULL,
  convert_lengths = TRUE,
  size_cutoff = NULL,
  zero_fill = TRUE,
  remove_unknown_lengths = TRUE,
  univariate = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{sources}{Which data sources would you like included? Options include: "Baystudy", "Suisun", "FMWT", "SKT", "DJFMP", "EDSM", and "TMM".}

\item{species}{Character vector of species to include. Set \code{species=NULL} (the default) to include all species. It is recommended to filter to species of interest using this option.}

\item{convert_lengths}{Should all species with conversion equations have their lengths converted to fork length (or total length if no fork)?}

\item{size_cutoff}{Minimum fish length (in mm) that should be retained in the data. Any fish below this length will be removed.
It is advised to convert lengths before using this option. All \code{NA} lengths will be retained. Set \code{size_cutoff=NULL} to keep all sizes.}

\item{zero_fill}{Should zeros be filled in for samples in which a species was not recorded as caught, using the \code{\link{zero_fill}} function. The \code{remove_unknown_lengths} and \code{univariate} parameters control the behavior of this functionality and these parameters are passed to the \code{\link{zero_fill}} function.}

\item{remove_unknown_lengths}{Should samples associated with unknown lengths be removed from the data? Unknown lengths refer
to rows with \code{Length_NA_flag=="Unknown length"}, corresponding to Suisun samples where unmeasured fish were not a random
sample from the same pool fish were selected to be measured, AND the length range could not be estimated from the comments, or
no fish within that size range were measured in that sample. The behavior of this option depends on the \code{univariate} parameter.}

\item{univariate}{Controls behavior of the \code{remove_unknown_lengths} option. If \code{remove_unknown_lengths=FALSE} this parameter is ignored.
Will these data be used for univariate analyses (\code{univariate=TRUE})? Or multi-species analyses (\code{univariate=FALSE})?
If univariate, when a \code{Length_NA_flag=="Unknown length"} record is found, all records of that taxa from that sample are removed and no 0s are filled in.
In effect, this is transforming those records into missing data. If \code{univariate=FALSE}, when a \code{Length_NA_flag=="Unknown length"} record is found,
the entire sample is removed and no 0s are filled in, since accurate community data cannot be confirmed for that sample.}

\item{quiet}{Set \code{quiet=TRUE} to suppress the warning message about length data.}
}
\description{
Integrates your desired fish datasets according to a number of user-specified parameters
}
\details{
It is recommended to use the \code{species} parameter to filter the data to the species (singular or plural) of interest, otherwise it will create a gigantic dataset.
You can fill in 0s for multiple univariate analyses simultaneously by passing the species of interest to the \code{species} parameter and setting \code{univariate=TRUE}.
}
\examples{
\dontrun{
data<-fish(c("Baystudy", "Suisun", "FMWT", "SKT", "DJFMP", "EDSM"),
size_cutoff=40)
}
}
\seealso{
\code{\link{Suisun}}, \code{\link{Suisun_measured_lengths}}, \code{\link{FMWT}}, \code{\link{FMWT_measured_lengths}}, \code{\link{Baystudy}}, \code{\link{Baystudy_measured_lengths}}
}
