% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FMWT_measured_lengths}
\alias{FMWT_measured_lengths}
\title{FMWT measured lengths}
\format{
a tibble with 240,495 rows and 4 variables
\describe{
\item{SampleID}{Unique sample identifier.}
\item{Taxa}{Scientific name.}
\item{Length}{Sampling survey (loosely corresponds to month).}
\item{Count}{Number of fish measured to this length.}
}
}
\usage{
FMWT_measured_lengths
}
\description{
Only measured lengths from the FMWT data
}
\keyword{datasets}
