% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Length_conversions}
\alias{Length_conversions}
\title{Length conversions}
\format{
a tibble with 20 rows and 3 variables
\describe{
\item{Species}{Scientific name.}
\item{Intercept}{Intercept in the equation FL = Intercept + Slope * SL.}
\item{Slope}{Slope in the equation FL = Intercept + Slope * SL.}
}
}
\usage{
Length_conversions
}
\description{
Intercepts and slopes for equations to convert Standard Length to Fork Length (or Total Length if no fork). Equations derived from the CDFW length-weight study (2005) and Jereme Gaeta, unpublished.
}
\keyword{datasets}
