% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLSIntegrate.R
\name{SLSIntegrate}
\alias{SLSIntegrate}
\title{Integration of SLS Access base tables into a long format data frame}
\usage{
SLSIntegrate(filePath = NULL)
}
\arguments{
\item{filePath}{Location of the base Access tables to be integrated. Default behavior is to simply read from the file path within data-raw/SLS folder of this package.}
}
\value{
An integrated fish dataset of the CDFW SLS Survey.
}
\description{
This function will read in the individual base tables from the SLS Access
database and integrate them into a singular long data format. The required
tables are "Catch", "Lengths", "Tow Info", "Water Info", and "20 mm Stations"
from the Access database.
}
