% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_fill.R
\name{zero_fill}
\alias{zero_fill}
\title{Fill in 0 catches}
\usage{
zero_fill(
  data,
  species = NULL,
  remove_unknown_lengths = TRUE,
  univariate = TRUE
)
}
\arguments{
\item{data}{The input data, likely the output from \code{\link{LTMRpilot}}, \code{\link{fish}}, or some row-bound combination of the internal package datasets.
Must contain at least the SampleID, Taxa, Length_NA_flag (if \code{remove_unmeasured_lengths=TRUE}), and Count columns.}

\item{species}{Character vector of species to include. Set \code{species=NULL} (the default) to include all species. It is recommended to filter to species of interest using this option.}

\item{remove_unknown_lengths}{Should samples associated with unknown lengths be removed from the data? Unknown lengths refer
to rows with \code{Length_NA_flag=="Unknown length"}, corresponding to Suisun samples where unmeasured fish were not a random
sample from the same pool fish were selected to be measured, AND the length range could not be estimated from the comments, or
no fish within that size range were measured in that sample. The behavior of this option depends on the \code{univariate} parameter.}

\item{univariate}{Controls behavior of the \code{remove_unknown_lengths} option. If \code{remove_unknown_lengths=FALSE} this parameter is ignored.
Will these data be used for univariate analyses (\code{univariate=TRUE})? Or multi-species analyses (\code{univariate=FALSE})?
If univariate, when a \code{Length_NA_flag=="Unknown length"} record is found, all records of that taxa from that sample are removed and no 0s are filled in.
In effect, this is transforming those records into missing data. If \code{univariate=FALSE}, when a \code{Length_NA_flag=="Unknown length"} record is found,
the entire sample is removed and no 0s are filled in, since accurate community data cannot be confirmed for that sample.}
}
\description{
Fill in 0s for samples in which a species was not recorded as caught.
}
\details{
It is recommended to use the \code{species} parameter to filter the data to the species (singular or plural) of interest, otherwise it will create a gigantic dataset.
You can fill in 0s for multiple univariate analyses simultaneously by passing the species of interest to the \code{species} parameter and setting \code{univariate=TRUE}.
}
\examples{

# Using all species with converted lengths
Data <- LTMRpilot(convert_lengths=TRUE, remove_unconverted_lengths=TRUE)

# For a univariate analysis
species <- c("Clupea pallasii", "Morone saxatilis", "Parophrys vetulus", "Sardinops sagax")
Data_filled <- zero_fill(Data, species=species, remove_unknown_lengths=TRUE, univariate=TRUE)

# For a multivariate analysis
Data_filled <- zero_fill(Data, remove_unknown_lengths=TRUE, univariate=FALSE)

}
\seealso{
\code{\link{LTMRpilot}}
}
