% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Baystudy}
\alias{Baystudy}
\title{Bay study dataset}
\format{
a tibble with 743,409 rows and 35 variables
\describe{
  \item{Tow_volume}{Volume towed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}}).}
  \item{Tow_area}{Area towed (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^{2}}}).}
  \item{Latitude}{May not be accurate, see details.}
  \item{Longitude}{May not be accurate, see details.}
  }
  @details Some station locations have moved over time due to shoals filling in, trees, etc.
}
\usage{
Baystudy
}
\description{
California Department of Fish and Wildlife Bay Study data.
}
\keyword{datasets}
