% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Baystudy}
\alias{Baystudy}
\title{Bay study dataset}
\format{
a tibble with xxx rows and xx variables
\describe{
  \item{Latitude}{May not be accurate, see details.}
  \item{Longitude}{May not be accurate, see details.}
  }
  @details Some station locations have moved over time due to shoals filling in, trees, etc.
}
\usage{
Baystudy
}
\description{
California Department of Fish and Wildlife Bay Study data.
}
\keyword{datasets}
