% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Baystudy_measured_lengths}
\alias{Baystudy_measured_lengths}
\title{Baystudy measured lengths}
\format{
a tibble with 864,599 rows and 5 variables
\describe{
\item{SampleID}{Unique sample identifier.}
\item{Taxa}{Scientific name.}
\item{Size_group}{When two different size groups of a species are collected, Bay study may split them into size groups and subsample from each size group separately for fish to measure.}
\item{Length}{Fork length, total length if there's no fork or heterocercal tail (sturgeon, shark). Bat Ray wing widths starting around 1989-1992 to present. (mm).}
\item{Count}{Number of fish measured to this length.}
}
}
\usage{
Baystudy_measured_lengths
}
\description{
Only measured lengths from the Baystudy data
}
\keyword{datasets}
