% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FMWT}
\alias{FMWT}
\title{FMWT dataset}
\format{
a tibble with 220,785 rows and 35 variables
\describe{
  \item{Tow_volume}{Volume towed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}}).}
  }
}
\usage{
FMWT
}
\description{
California Department of Fish and Wildlife Fall Midwater Trawl data.
}
\details{
More metadata and information on methods are available \href{https://www.dfg.ca.gov/delta/projects.asp?ProjectID=FMWT}{here}.
}
\keyword{datasets}
