% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRdata.R
\docType{package}
\name{LTMRdata}
\alias{LTMRdata-package}
\alias{LTMRdata}
\title{LTMRdata: A package for integrating long-term monitoring datasets for the IEP long-term monitoring program review.}
\description{
This package contains internal datasets and a function to bind them all together.
}
\section{Data}{

\itemize{
  \item \code{\link{Suisun}}
  \item \code{\link{FMWT}}
  \item \code{\link{Baystudy}}
  \item \code{\link{TMM}}
  \item \code{\link{Species}}
}
}

\section{Functions}{

\itemize{
  \item \code{\link{LTMRpilot}}
  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbashevkin/LTMRdata}
  \item Report bugs at \url{https://github.com/sbashevkin/LTMRdata/issues}
}

}
\author{
\strong{Maintainer}: Sam Bashevkin \email{sam.bashevkin@waterboards.ca.gov} (\href{https://orcid.org/0000-0001-7406-7089}{ORCID})

}
