% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRdata.R
\docType{package}
\name{LTMRdata}
\alias{LTMRdata}
\title{LTMRdata: A package for integrating long-term monitoring datasets for the IEP long-term monitoring program review.}
\description{
This package contains internal datasets and a function to bind them all together.
}
\section{Data}{

\itemize{
  \item \code{\link{Suisun}}
  \item \code{\link{FMWT}}
  \item \code{\link{Baystudy}}
}
}

\section{Functions}{

\itemize{
  \item \code{\link{LTMRpilot}}
  }
}

