% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRpilot.R
\name{LTMRpilot}
\alias{LTMRpilot}
\title{Integrate datasets for pilot review}
\usage{
LTMRpilot()
}
\description{
Integrate FMWT, Suisun, and Bay study datasets for IEP long-term monitoring program pilot review
}
\examples{
Data <- LTMRpilot()
}
\seealso{
\code{\link{Suisun}}, \code{\link{FMWT}}, \code{\link{Baystudy}}
}
