% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRpilot.R
\name{LTMRpilot}
\alias{LTMRpilot}
\title{Integrate datasets for pilot review}
\usage{
LTMRpilot(quiet = FALSE, measured_lengths = FALSE)
}
\arguments{
\item{quiet}{Set \code{quiet=TRUE} to suppress the warning message about length data.}

\item{measured_lengths}{Only include measured lengths. Does not calculate adjusted length frequencies. Will not represent total fish caught.}
}
\description{
Integrate FMWT, Suisun, and Bay study datasets for IEP long-term monitoring program pilot review
}
\examples{
Data <- LTMRpilot()

#Only outputting measured lengths, without frequencies adjusted for counts of unmeasured fish
Data <- LTMRpilot(measured_lengths=TRUE)
}
\seealso{
\code{\link{Suisun}}, \code{\link{Suisun_measured_lengths}}, \code{\link{FMWT}}, \code{\link{FMWT_measured_lengths}}, \code{\link{Baystudy}}, \code{\link{Baystudy_measured_lengths}}
}
