% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{Salvage}
\alias{Salvage}
\title{Salvage dataset}
\format{
a tibble with 2,514,681 rows and 20 variables
\describe{
  \item{Source}{Name of source dataset.}
  \item{Station}{State or federal facility buildings (of the SWP or CVP).}
  \item{Latitude}{Latitude (decimal degrees) of the full facility.}
  \item{Longitude}{Longitude (decimal degrees) of the full facility.}
  \item{Date}{Date sample was collected.}
  \item{Datetime}{Date and time sample was collected.}
  \item{SampleID}{Unique sample identifier.}
  \item{Method}{Sampling period (0000 = Normal count, 9999 = Second flush)}
  \item{Tow_volume}{The amount of water export in cubic meter.}
  \item{Temp_surf}{Surface temperature in °C.}
  \item{Taxa}{Scientific name.}
  \item{Length}{Fork length from tip of the snout to a point at the fork of the caudal fin (mm) or total length for species without a forked tail.}
  \item{Count}{Estimated count for each sample, taxa, and length.}
  \item{Notes_tow}{Comments or notes of the sampling period.}
  \item{Length_NA_flag}{Why is the length NA?}
  }
}
\usage{
Salvage
}
\description{
California Department of Fish and Wildlife Fish Salvage dataset.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Salvage-Monitoring}{here}.
}
\keyword{datasets}
