% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{SKT_measured_lengths}
\alias{SKT_measured_lengths}
\title{SKT dataset}
\format{
a tibble with X rows and Y number of variables
\describe{
  \item{SampleID}{Unique sample identifier.}
  \item{Taxa}{Scientific name.}
  \item{Length}{Fork length, total length if there's no fork or heterocercal tail (sturgeon, shark).}
  \item{Count}{Estimated count for each sample, taxa, and length.}
  }
}
\usage{
SKT_measured_lengths
}
\description{
California Department of Fish and Wildlife Spring Kodiac Trawl (SKT) data.
}
\keyword{datasets}
