% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{SLS}
\alias{SLS}
\title{SLS dataset}
\format{
a tibble with `r prettyNum(nrow(SLS), big.mark = ",")` rows and `r ncol(SLS)` variables
\describe{
  \item{Source}{Name of source dataset.}
  \item{Station}{Station where sample was collected.}
  \item{Latitude}{Latitude (decimal degrees).}
  \item{Longitude}{Longitude (decimal degrees).}
  \item{Date}{Date sample was collected.}
  \item{Datetime}{Date and time sample was collected.}
  \item{Survey}{Survey number, roughly corresponding to month.}
  \item{Depth}{Bottom depth (m).}
  \item{SampleID}{Unique sample identifier.}
  \item{Method}{Sampling method.}
  \item{Tide}{Tidal stage.}
  \item{Sal_surf}{Surface salinity.}
  \item{Sal_bot}{Bottom salinity.}
  \item{Temp_surf}{Surface temperature in °C.}
  \item{Secchi}{Secchi depth (cm).}
  \item{TurbidityNTU}{Turbidity (NTU).}
  \item{TurbidityFNU}{Turbidity (FNU).}
  \item{Tow_volume}{Volume towed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}}).}
  \item{Cable_length}{Length of cable released when net deployed (m).}
  \item{Tow_duration}{Duration of tow (minutes).}
  \item{Taxa}{Scientific name.}
  \item{Length}{Fork length (mm).}
  \item{Count}{Estimated count for each sample, taxa, and length.}
  \item{Length_NA_flag}{Why is the length NA?}
  \item{Notes_tow}{Notes or comments on the trawl.}
  \item{Notes_flowmeter}{Notes or comments on the flowmeter reading.}
  }
}
\usage{
SLS
}
\description{
California Department of Fish and Wildlife Smelt Larva Survey data.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Smelt-Larva-Survey}{here}.
}
\keyword{datasets}
