% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species}
\alias{Species}
\title{Species codes}
\format{
a tibble with 322 rows and 13 variables
\describe{
\item{ScientificName}{Scientific Name.}
\item{Baystudy_Code}{Bay Study code.}
\item{CommonName}{Common name.}
\item{USFWS_Code}{EDSM and DJFMP codes.}
\item{SKT_Code}{Spring Kodiak Trawl code.}
\item{STN_Code}{Summer Townet Survey code.}
\item{SMF_Code}{Suisun Marsh Fish Study code.}
\item{FMWT_Code}{Fall Midwater Trawl code.}
\item{TMM_Code}{20-mm Survey code.}
\item{SLS_Code}{SLS Survey code.}
\item{Salvage_Code}{Salvage Survey code.}
\item{Lifestage}{Lifestage if specified.}
\item{Taxa}{Scientific name and life stage.}
}
}
\usage{
Species
}
\description{
Crosswalk table of species codes to common and scientific names
}
\keyword{datasets}
