% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species}
\alias{Species}
\title{Species codes}
\format{
a tibble with 299 rows and 7 variables
\describe{
  \item{Baystudy_Code}{Bay Study code.}
  \item{CommonName}{Common name.}
  \item{SMF_Code}{Suisun Marsh Fish Study code.}
  \item{FMWT_Code}{Fall Midwater Trawl code.}
  \item{ScientificName}{Scientific name.}
  \item{Lifestage}{Lifestage if specified.}
  \item{Taxa}{Combination of scientific name and life stage.}
  }
}
\usage{
Species
}
\description{
Crosswalk table of species codes to common and scientific names
}
\keyword{datasets}
