% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species}
\alias{Species}
\title{Species codes}
\format{
a tibble with 254 rows and 11 variables
\describe{
  \item{Baystudy_Code}{Bay Study code.}
  \item{CommonName}{Common name.}
  \item{SMF_Code}{Suisun Marsh Fish Study code.}
  \item{SKT_Code}{Spring Kodiak Trawl code.}
  \item{FMWT_Code}{Fall Midwater Trawl code.}
  \item{TMM_Code}{20-mm Survey code.}
  \item{STN_Code}{Summer Townet Survey code.}
  \item{ScientificName}{Scientific name.}
  \item{Taxa}{Scientific name and life stage.}
  \item{Lifestage}{Lifestage if specified.}
  \item{USFWS_Code}{EDSM and DJFMP codes.}
  }
}
\usage{
Species
}
\description{
Crosswalk table of species codes to common and scientific names
}
\keyword{datasets}
