% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{STN}
\alias{STN}
\title{STN dataset}
\format{
a tibble with 198,359 rows and 22 variables
\describe{
\item{Source}{Name of source dataset.}
\item{Station}{Station where sample was collected.}
\item{Latitude}{Latitude (decimal degrees).}
\item{Longitude}{Longitude (decimal degrees).}
\item{Date}{Date sample was collected.}
\item{Datetime}{Date and time sample was collected.}
\item{Survey}{Survey number, roughly corresponding to month.}
\item{Depth}{Bottom depth (m).}
\item{SampleID}{Unique sample identifier.}
\item{Method}{Sampling method (STN trawl).}
\item{Tide}{Tidal stage.}
\item{Sal_surf}{Surface salinity.}
\item{Temp_surf}{Surface temperature in °C.}
\item{TurbidityNTU}{Turbidity (NTU).}
\item{Secchi}{Secchi depth (cm).}
\item{Tow_volume}{Volume towed (\ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}}).}
\item{Tow_direction}{Tow direction relative to current.}
\item{Cable_length}{Length of cable released when net deployed (feet).}
\item{Taxa}{Scientific name.}
\item{Length}{Fork length from the most anterior part of the fish to the median caudal fin rays (mm).}
\item{Count}{Estimated count for each sample, taxa, and length.}
\item{Length_NA_flag}{Why is the length NA?}
}
}
\usage{
STN
}
\description{
California Department of Fish and Wildlife Summer Townet Survey (STN) data.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Townet-Survey}{here}.
}
\keyword{datasets}
