% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STN_measured_lengths}
\alias{STN_measured_lengths}
\title{STN measured lengths}
\format{
a tibble with `r prettyNum(nrow(STN_measured_lengths), big.mark = ",")` rows and `r ncol(STN_measured_lengths)` variables
\describe{
  \item{SampleID}{Unique sample identifier.}
  \item{Taxa}{Scientific name.}
  \item{Length}{Fork length from the most anterior part of the fish to the median caudal fin rays (mm).}
  \item{Count}{Number of fish measured to this length.}
  }
}
\usage{
STN_measured_lengths
}
\description{
Only measured lengths from the STN data
}
\keyword{datasets}
