% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Suisun}
\alias{Suisun}
\title{Suisun Marsh dataset}
\format{
a tibble with 155,446 rows and 19 variables
\describe{
  \item{Column}{Description.}
  }
}
\usage{
Suisun
}
\description{
UC Davis Suisun Marsh Fish Study data.
}
\keyword{datasets}
