% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Suisun_measured_lengths}
\alias{Suisun_measured_lengths}
\title{Suisun measured lengths}
\format{
a tibble with 167,785 rows and 5 variables
\describe{
  \item{SampleID}{Unique sample identifier.}
  \item{Taxa}{Scientific name.}
  \item{Dead}{Was fish dead?}
  \item{Length}{Standard length for all species except sturgeon, which are fork length (mm).}
  \item{Count}{Number of fish measured to this length.}
  }
}
\usage{
Suisun_measured_lengths
}
\description{
Only measured lengths from the Suisun data
}
\keyword{datasets}
