% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMM_measured_lengths}
\alias{TMM_measured_lengths}
\title{20mm measured lengths}
\format{
a tibble with 324,218 rows and 4 variables
\describe{
  \item{SampleID}{Unique sample identifier.}
  \item{Taxa}{Scientific name.}
  \item{Length}{Fork length from tip of the snout to a point at the fork of the caudal fin (mm).}
  \item{Count}{Number of fish measured to this length.}
  }
}
\usage{
TMM_measured_lengths
}
\description{
Only measued lengths from the 20-mm Survey data
}
\keyword{datasets}
